refactored ...move to separate class
