+++ src/NewBaghbondi/GameAudio.java
@@ -3,18 +3,29 @@ package NewBaghbondi;
 
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
+import javafx.util.Duration;
+
 import java.io.File;
 import java.nio.file.Paths;
 
 public class GameAudio {
-    MediaPlayer mediaPlayer;
-    public void introAudio(){
-        String path="resources/gameIntroMusic.mp3";
-//      Media media=new Media(new File(path).toURI().toString());
-        Media media=new Media(Paths.get(path).toUri().toString());
-        mediaPlayer=new MediaPlayer(media);
-        mediaPlayer.play();
 
+        public void introAudio(int status){
+            String path="src/resources/gameIntroMusic.mp3";
+            Media media=new Media(new File(path).toURI().toString());
+            MediaPlayer mediaPlayer=new MediaPlayer(media);
+            if(status==0) {
+                mediaPlayer.setOnEndOfMedia(new Runnable() {
+                    @Override
+                    public void run() {
+                        mediaPlayer.seek(Duration.ZERO);
+                    }
+                });
+                mediaPlayer.play();
+            }
+            else if(status==1){
+                if(mediaPlayer.getStatus()!=null)mediaPlayer.stop();
+            }
+        }
     }
 
-}
